alar_test/requirements.txt
django==3.1
psycopg2==2.8.5
gunicorn==20.0.4
celery==4.4.7
redis==3.5.3
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0